requirements.txt
google-generativeai==0.5.2
python-dotenv==1.0.1
